requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2.2
django-crontab==0.7.1
djangorestframework==3.14.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
mypy-extensions==1.0.0
outcome==1.2.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
psycopg2-binary==2.9.6
PySocks==1.7.1
pytz==2023.3
requests==2.31.0
schedule==1.2.0
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.0
trio-websocket==0.10.3
urllib3==2.0.3
wsproto==1.2.0
